+++ lib/src/bones_api_base.dart
@@ -32,7 +32,7 @@ typedef APILogger = void Function(APIRoot apiRoot, String type, String? message,
 
 class BonesAPI {
   // ignore: constant_identifier_names
-  static const String VERSION = '1.2.23';
+  static const String VERSION = '1.2.24';
 
   static bool _boot = false;